scripts/vault_transform/requirements.txt
pyyaml>=6.0.2